+++ src/languages/es.js
@@ -1,6 +1,0 @@
-/* eslint-disable quotes */
-Lang.prototype.pack.es = {
-    token: {
-        // TODO
-    },
-};