+++ src/languages/ru.js
@@ -1,6 +1,0 @@
-/* eslint-disable quotes */
-Lang.prototype.pack.ru = {
-    token: {
-        // TODO
-    },
-};